refactor: Move BuildingBlockSagas to separate directory
